# api/docs/openapi.yaml
openapi: 3.0.3
paths:
...
  /api/v1/users/me/categories/{categoryId}/rules:
    put:
      description: Legacy personal compatibility route. Atomic replacement resolves the authenticated User's Personal Space; when a revision is supplied, stale replacements are rejected. Unchanged normalized pattern/match type pairs retain IDs and createdAt; unchanged display patterns also retain updatedAt. Empty rules clears the set. Returns ascending ID order. Duplicate normalized pattern/match type pairs within the request or another Personal Space Category return 409 with the conflicting categoryId and request field. Inactive Categories return 409, including empty replacements. Validation and conflict failures change nothing.
      operationId: CategoryRuleReplacement_replaceCategoryRules
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceCategoryRulesDto'
      responses:
        '200':
          description: Persisted rules for the selected Category.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryRuleResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '413':
          $ref: '#/components/responses/HttpError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaTypeError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Replace the complete rule set for an active Category.
      tags:
        - Category rules
...
info:
  title: Expense Tracker REST API
  description: A user-scoped expense tracker API. Bigint identifiers and counts, decimal money, domain dates, and UTC timestamps use the encodings described by the schemas below.
  version: 1.0.0
  contact: {}
...
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - INVALID_JSON
        - NOT_ACCEPTABLE
        - UNSUPPORTED_MEDIA_TYPE
        - ROUTE_NOT_FOUND
        - HTTP_ERROR
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
        - UNAUTHENTICATED
        - USER_NOT_PROVISIONED
        - CONFLICT
        - STALE_EDIT
        - RESOURCE_NOT_FOUND
        - USER_NOT_FOUND
        - SPACE_NOT_FOUND
        - SPACE_NOT_WRITABLE
        - EMAIL_ALREADY_EXISTS
        - CATEGORY_NOT_FOUND
        - CATEGORY_NAME_ALREADY_EXISTS
        - CATEGORY_INACTIVE
        - BUDGET_NOT_FOUND
        - CATEGORY_RULE_NOT_FOUND
        - CATEGORY_RULE_PATTERN_ALREADY_EXISTS
        - TRANSACTION_NOT_FOUND
        - STATEMENT_IMPORT_NOT_FOUND
        - IMPORTED_TRANSACTION_IMMUTABLE
        - STATEMENT_IMPORT_FILE_ALREADY_EXISTS
        - STATEMENT_IMPORT_FILE_HASH_INVALID
        - STATEMENT_IMPORT_PROBABLE_DUPLICATES
      description: Stable error code. Human-readable messages are not a compatibility contract.
    ErrorResponseDto:
      type: object
      properties:
        code:
          description: Stable error code. Human-readable messages are not a compatibility contract.
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Human-readable error message.
        details:
          type: array
          description: Structured details for fields or domain-specific conflicts.
          items:
            oneOf:
              - $ref: '#/components/schemas/ErrorDetailDto'
              - $ref: '#/components/schemas/ProbableDuplicateDetailDto'
      additionalProperties: false
      required:
        - code
        - message
        - details
    ErrorEnvelopeDto:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorResponseDto'
      additionalProperties: false
      required:
        - error
    ErrorDetailCode:
      type: string
      enum:
        - required
        - invalid_format
        - invalid_type
        - invalid_value
        - unknown
        - incompatible
        - not_allowed
        - not_unique
        - too_short
        - too_long
        - probable_duplicate
      description: Stable machine-readable field-detail code.
    ErrorDetailDto:
      type: object
      properties:
        categoryId:
          type: string
          description: Conflicting owned Category identifier for a category-rule duplicate.
          pattern: ^[1-9]\d*$
          example: '42'
        field:
          type: string
          description: JSON Pointer into the request.
          example: /email
        code:
          description: Stable machine-readable field-detail code.
          allOf:
            - $ref: '#/components/schemas/ErrorDetailCode'
        message:
          type: string
          description: Human-readable detail message.
      additionalProperties: false
      required:
        - field
        - code
        - message
    ProbableDuplicateDetailDto:
      type: object
      properties:
        field:
          type: string
          example: /transactions/0
        code:
          type: string
          enum:
            - probable_duplicate
        message:
          type: string
        transactionIndexes:
          type: array
          items:
            type: integer
            minimum: 0
        committedTransactionIds:
          type: array
          description: Public transaction IDs encoded as strings.
          items:
            type: string
            pattern: ^[1-9]\d*$
      additionalProperties: false
      required:
        - field
        - code
        - message
        - transactionIndexes
        - committedTransactionIds
...
    CategoryRuleResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Positive bigint Category rule identifier encoded as a string.
          pattern: ^[1-9]\d*$
          example: '42'
        categoryId:
          type: string
          description: Positive bigint Category identifier encoded as a string.
          pattern: ^[1-9]\d*$
          example: '42'
        pattern:
          type: string
          description: Original rule text containing at least one non-whitespace character. Matching trims surrounding whitespace, collapses repeated whitespace, and compares case-insensitively.
          minLength: 1
          maxLength: 500
          pattern: \S
          example: Green   Market
        matchType:
          type: string
          description: Exact or Contains matching. The normalized pattern is not returned.
          enum:
            - exact
            - contains
          example: exact
        createdAt:
          type: string
          description: UTC timestamp when the Category rule was created.
          format: date-time
          example: '2026-08-29T00:00:00.000Z'
        updatedAt:
          type: string
          description: UTC timestamp when the Category rule was last updated.
          format: date-time
          example: '2026-08-29T00:00:00.000Z'
      description: A Category Rule belonging to the authorized Space. The original pattern is retained for display. Rule evaluation is performed by the frontend.
      additionalProperties: false
      required:
        - id
        - categoryId
        - pattern
        - matchType
        - createdAt
        - updatedAt
...
    ReplacementCategoryRuleDto:
      type: object
      properties:
        pattern:
          type: string
          minLength: 1
          maxLength: 500
          pattern: \S
        matchType:
          type: string
          enum:
            - exact
            - contains
      additionalProperties: false
      required:
        - pattern
        - matchType
    ReplaceCategoryRulesDto:
      type: object
      properties:
        revision:
          type: string
          description: The revision returned with the last Space-scoped rule collection. A stale replacement is rejected when another member changed any Rule first.
          pattern: ^\d+$
          example: '3'
        rules:
          description: Complete desired rule set. Empty removes all rules for this Category.
          type: array
          items:
            $ref: '#/components/schemas/ReplacementCategoryRuleDto'
      additionalProperties: false
      required:
        - rules
...
  responses:
    UnauthenticatedError:
      description: The request does not include valid Clerk session credentials.
      headers:
        WWW-Authenticate:
          required: true
          description: Authentication scheme accepted by the API.
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: UNAUTHENTICATED
              message: Authentication required
              details: []
    UserNotProvisionedError:
      description: The verified Clerk identity is not linked to a local User. Provision the User first.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: USER_NOT_PROVISIONED
              message: User is not provisioned
              details: []
    NotFoundError:
      description: The requested resource is absent or owned by another user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            user:
              value:
                error:
                  code: USER_NOT_FOUND
                  message: User was not found
                  details: []
            route:
              value:
                error:
                  code: ROUTE_NOT_FOUND
                  message: Request failed
                  details: []
    NotAcceptableError:
      description: The client did not accept the representation required by the endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: NOT_ACCEPTABLE
              message: The endpoint requires an application/json response
              details: []
    InternalError:
      description: An unexpected server error. Infrastructure details are never returned.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: INTERNAL_ERROR
              message: An internal error occurred
              details: []
    ValidationError:
      description: The request syntax or fields are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            invalidField:
              value:
                error:
                  code: VALIDATION_FAILED
                  message: The request contains invalid fields.
                  details:
                    - field: /email
                      code: invalid_format
                      message: Email must be a valid email address
    ConflictError:
      description: The request conflicts with current application state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            emailAlreadyExists:
              value:
                error:
                  code: EMAIL_ALREADY_EXISTS
                  message: Email is already in use
                  details:
                    - field: /email
                      code: not_unique
                      message: Email is already in use
            genericConflict:
              value:
                error:
                  code: CONFLICT
                  message: Request conflicts with existing data
                  details: []
...
    HttpError:
      description: A generic HTTP protocol error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: HTTP_ERROR
              message: Request failed
              details: []
    UnsupportedMediaTypeError:
      description: A request body must use application/json.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: UNSUPPORTED_MEDIA_TYPE
              message: Request bodies must use application/json
              details: []
...
security:
  - bearerAuth: []